move convertTickProps to util.ts
